pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "support-bot"
version = "0.1.0"
description = "Orbio AI — Intelligent Conversational Support Agent built with Google ADK"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.0.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    # Voice mode: real-time audio I/O for ADK run_live() bidirectional streaming.
    # google-cloud-speech and google-cloud-texttospeech are NOT needed —
    # gemini-live-2.5-flash-native-audio handles STT+TTS natively.
    "sounddevice>=0.4",
    "numpy>=1.24",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["support_bot"]
